Moved check_rate_limit to utils
